Finished with geometry.ts breakup
